uni_gofundme/requirements.txt
asgiref==3.2.3
boto3==1.12.20
botocore==1.15.20
Django==3.0.3
django-cors-headers==3.2.1
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
jmespath==0.9.5
Markdown==3.2.1
Pillow==7.0.0
postgres==3.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
psycopg2-pool==1.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
